[feat] 기존 StackView(CategoryView + LogoView) 를 MenuCategoryViewController에서, MenuCategoryStackView 파일을 생성하여 분리
